% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{calculate_stats_light}
\alias{calculate_stats_light}
\title{Calculate statistics for a core clustering experiment}
\usage{
calculate_stats_light(res, trimmed_core = FALSE)
}
\arguments{
\item{res}{A results structure.}

\item{trimmed_core}{Boolean indicating whether the clustering algorithm used trimming.}
}
\value{
A matrix with statistics.
}
\description{
Calculate statistics for a core clustering experiment
}

