% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_clustering_functions.R
\name{cluster_hclust}
\alias{cluster_hclust}
\title{Return cluster indices using hierarchical clustering.}
\usage{
cluster_hclust(x, k)
}
\arguments{
\item{x}{Data matrix}

\item{n}{Number of clusters.}
}
\value{
Vector with cluster labels.
}
\description{
Return cluster indices using hierarchical clustering.
}

