% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_clustering_functions.R
\name{cluster_kmeanspp}
\alias{cluster_kmeanspp}
\title{Return cluster indices using kmeans++.}
\usage{
cluster_kmeanspp(x, k, iteration = NULL)
}
\arguments{
\item{x}{Data matrix}

\item{n}{Number of clusters}
}
\value{
Vector with cluster labels.
}
\description{
Return cluster indices using kmeans++.
}

