% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_clustering_functions.R
\name{cluster_tclust}
\alias{cluster_tclust}
\title{Cluster a dataset x into k clusters using tclust}
\usage{
cluster_tclust(x, k)
}
\arguments{
\item{x}{The data}

\item{k}{The number of clusters}
}
\value{
The cluster indices of the data.
}
\description{
Cluster a dataset x into k clusters using tclust
}

