% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_clustering_functions.R
\name{clusterfunc_svm}
\alias{clusterfunc_svm}
\title{Cluster (classify) a dataset x into its constituent clusters (classes)
using an SVM.}
\usage{
clusterfunc_svm(x)
}
\arguments{
\item{x}{The data}
}
\value{
The cluster indices of the data.
}
\description{
Cluster (classify) a dataset x into its constituent clusters (classes)
using an SVM.
}

