% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_corecluster.R
\name{clusters_from_cliques}
\alias{clusters_from_cliques}
\title{Get the clusters from the cliques}
\usage{
clusters_from_cliques(cliques, N, weak_cluster = 0)
}
\arguments{
\item{cliques}{The cliques}

\item{N}{The number of data items.}

\item{weak_cluster}{The id of the weak points. The default is 0.}
}
\value{
The clusters
}
\description{
Get the clusters from the cliques
}

