% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_corecluster.R
\name{co_occurrence}
\alias{co_occurrence}
\title{A helper function for that determines if two data items x1 and x2 occur in one
clustering of the data}
\usage{
co_occurrence(x1, x2, data_function, cluster_function)
}
\arguments{
\item{x1}{Data item i.}

\item{x2}{Data item j.}

\item{data_function}{A function, such that data_function() returns n-2 resampled data points.}

\item{cluster_function}{A clustering function that returns a set of cluster labels.}
}
\value{
1 if x1 and x2 are in the same clustering, 0 if not. NULL if the clustering failed.
}
\description{
A helper function for that determines if two data items x1 and x2 occur in one
clustering of the data
}

