% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_corecluster.R
\name{co_occurrence_prob_all}
\alias{co_occurrence_prob_all}
\title{Estimate all co-occurence probabilities Pij using bootstrap and within-bag-samples}
\usage{
co_occurrence_prob_all(data, cluster_function, iter = 1000)
}
\arguments{
\item{data}{A matrix with data points.}

\item{cluster_function}{A clustering function that returns a set of cluster labels.}

\item{iter}{The number of iterations. Optional. Default is 1000.}
}
\value{
A square matrix of size (nrow(data) x nrow(data))
containing the co-occurence probabilities of the items in the data
matrix.
}
\description{
Estimate all co-occurence probabilities Pij using bootstrap and within-bag-samples
}

