% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{confusion_matrix_to_latex}
\alias{confusion_matrix_to_latex}
\title{Given a list of algorithm names and two lists with results,
create a latex table.}
\usage{
confusion_matrix_to_latex(alg_list, list_left, list_right)
}
\arguments{
\item{alg_list}{A list with the names of clustering algorithms used in the experiments}

\item{list_left}{The contents for the left side of the list}

\item{list_right}{The contents for the rigth side of the list}
}
\value{
Outputs a latex table
}
\description{
Given a list of algorithm names and two lists with results,
create a latex table.
}

