% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_corecluster.R
\name{core_clustering}
\alias{core_clustering}
\title{Core clustering.}
\usage{
core_clustering(data, clustering_func = NULL, sampling_func = NULL,
  method = "bootstrap", alpha = 0.1, n_iter = 1000, weak_cluster = -1,
  using_trim = FALSE, seed = NULL)
}
\arguments{
\item{data}{The data. Either a matrix or a data frame, depending on the requirements of the clustering function.}

\item{clustering_func}{A clustering function to use, which given data returns cluster indices.}

\item{sampling_func}{A sampling function to use. Default is NULL. Only used when method is \code{distribution}.}

\item{method}{The method to use. String. Either \code{bootstrap} to
use bootstrapping or \code{distribution} to calculate the exact
co-occurrence probabilities.}

\item{alpha}{The confidenence level. Default 0.1}

\item{n_iter}{The number of iterations to use. The default is 1000.}

\item{weak_cluster}{The idnex to use for weak points. Default is -1.}

\item{using_trim}{Boolean. Does the clustering method use trimming?}

\item{seed}{Random seed. Default is NULL to not set the random seed.}
}
\value{
A list structure:
\describe{
\item{data}{The data}
\item{alpha}{The confidence level}
\item{iterations}{The number of iterations}
\item{cluster}{The initial clustering}
\item{cluster_core}{The core clustering}
\item{mat}{The co-occurrence probability matrix}
\item{amat}{The adjacency matrix}
\item{using_trim}{Whether or not the clustering method is using trimming}
\item{weak_cluster}{The index used to describe the weak clusters}
}
}
\description{
Core clustering.
}

