% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{dataset_specs}
\alias{dataset_specs}
\title{Get the specifications of a dataset}
\usage{
dataset_specs(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
A list with the specifications of the dataset
}
\description{
Get the specifications of a dataset
}

