% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_corecluster.R
\name{find_core_clusters}
\alias{find_core_clusters}
\title{Determine the core clusters}
\usage{
find_core_clusters(g, clusters)
}
\arguments{
\item{g}{A graph}

\item{clusters}{An initial clustering}
}
\value{
The core clusters
}
\description{
Determine the core clusters
}

