% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_corecluster.R
\name{find_largest_clique}
\alias{find_largest_clique}
\title{Find the largest clique in the graph g}
\usage{
find_largest_clique(g)
}
\arguments{
\item{g}{A graph}
}
\value{
The indices of the items in the largest clique in g.
}
\description{
Find the largest clique in the graph g
}

