% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{fix_alg_name}
\alias{fix_alg_name}
\title{Fix the names of clustering algorithms}
\usage{
fix_alg_name(x)
}
\arguments{
\item{x}{The name of a clustering algorithm as a string}
}
\value{
A string
}
\description{
Fix the names of clustering algorithms
}
\keyword{internal}

