% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{fix_clusters}
\alias{fix_clusters}
\title{Match clusterings given a template using the Hungarian algorithm}
\usage{
fix_clusters(res_template, res_other)
}
\arguments{
\item{res_template}{A result structure to be used as a template}

\item{res_other}{Another result structure, which is to be modified}
}
\value{
res_other with the clusters matched to res_template
}
\description{
Match clusterings given a template using the Hungarian algorithm
}

