% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{get_agreement}
\alias{get_agreement}
\title{Get the running time of a core clustering experiment}
\usage{
get_agreement(dir_1, dir_2, ds_name, alg_name)
}
\arguments{
\item{dir_1}{A directory with results}

\item{dir_2}{A directory with results to compare to dir_1}

\item{ds_name}{The name of the dataset}

\item{alg_name}{The name of the clustering algorithm}
}
\value{
A list with confusion matrices for the agreement between
the results
}
\description{
Get the running time of a core clustering experiment
}

