% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{get_confusion_matrix}
\alias{get_confusion_matrix}
\title{Make a confusion matrix of the core clustering agreement
for clusters cl_1 and cl_2}
\usage{
get_confusion_matrix(cl_1, cl_2)
}
\arguments{
\item{cl_1}{A vector with cluster indices}

\item{cl_2}{A vector with cluster indices}
}
\value{
A confusion matrix
}
\description{
Make a confusion matrix of the core clustering agreement
for clusters cl_1 and cl_2
}

