% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_data.R
\name{get_data_kdd}
\alias{get_data_kdd}
\title{Load the KDD Cup 1999 dataset}
\usage{
get_data_kdd(kdd_data_path, N_samples = 200, format = "df", mode = "full",
  seed = 42)
}
\arguments{
\item{kdd_data_path}{The path to the data}

\item{N_samples}{The number of samples to obtain. Default is 200.}

\item{format}{String. Either \code{df} to return a data frame or \code{matrix} to return a matrix.}

\item{mode}{either \code{sample} to return N_samples of data items or \code{full} (default) to return the entire dataset}

\item{seed}{Random seed used when mode is \code{sample}. Default is 42.}
}
\value{
The data
}
\description{
Load the KDD Cup 1999 dataset
}

