% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_data.R
\name{get_data_synthetic}
\alias{get_data_synthetic}
\title{Get synthetic data}
\usage{
get_data_synthetic(N = 150, data.mean = 0, data.sd = 1, type = "matrix")
}
\arguments{
\item{N}{The number of data points. Default is 150.}

\item{data.mean}{Data mean.}

\item{data.sd}{Data standard deviation.}

\item{type}{Type of data. String: 'matrix' or 'df'.}
}
\value{
A list with the synthetic data points and their corresponding classes (clusters).
}
\description{
Get synthetic data
}

