% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_data.R
\name{get_dataset_synthetic}
\alias{get_dataset_synthetic}
\title{Get the synthetic dataset}
\usage{
get_dataset_synthetic(N = 150, data.mean = 0, data.sd = 1, seed = 42,
  type = "matrix")
}
\arguments{
\item{N}{The number of data points. Default is 150.}

\item{data.mean}{Data mean.}

\item{data.sd}{Data standard deviation.}

\item{seed}{Random seed. Default is 42.}

\item{type}{Type of data. String: 'matrix' or 'df'.}
}
\value{
A list with the synthetic data points and their corresponding classes (clusters).
}
\description{
Get the synthetic dataset
}

