% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{get_reslist}
\alias{get_reslist}
\title{Read results from experiments and return them as a list}
\usage{
get_reslist(classifier_list, dataset, base_path = "results", is_list = TRUE,
  prefix = "res")
}
\arguments{
\item{classifier_list}{A list of classifier names}

\item{dataset}{The name of the dataset to read}

\item{base_path}{The base path to the results}

\item{is_list}{Is the result structure a list}

\item{prefix}{A prefix used in the filenames of the results}
}
\value{
A list of result structures from core clustering
}
\description{
Read results from experiments and return them as a list
}

