% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{get_running_time}
\alias{get_running_time}
\title{Get the running time of a core clustering experiment}
\usage{
get_running_time(res)
}
\arguments{
\item{res}{A result structure from a core clustering experiment.}
}
\value{
The running time for the experiment
}
\description{
Get the running time of a core clustering experiment
}

