% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plots.R
\name{get_symbol_vector}
\alias{get_symbol_vector}
\title{Get a list of symbols given the clusters.}
\usage{
get_symbol_vector(clusters, symbol_list)
}
\arguments{
\item{clusters}{The clusters}

\item{symbol_list}{A list of plot symbols.}
}
\value{
A list of symbols
}
\description{
Get a list of symbols given the clusters.
}

