% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_clustering_functions.R
\name{kmeanspp}
\alias{kmeanspp}
\title{K-means++ algorithm.}
\usage{
kmeanspp(x, k, iter.max = 1000, algorithm = "Lloyd")
}
\arguments{
\item{x}{Data matrix with rows corresponding to data items.}

\item{k}{Number of clusters.}

\item{algorithm}{Clustering algorithm to use. Optional. Default is
\code{Lloyd}. See documentation for \code{kmeans} for details.}

\item{iter}{The maximum number of iterations. Optional. Default is 1000.}
}
\value{
See documentation for kmeans for details.
}
\description{
K-means++ algorithm.
}
\references{
Arthur & Vassilivitskii (2007) k-means++: the
advantages of careful seeding. In Proc SODA '07, 1027-1035.
}

