% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_clustering_functions.R
\name{kmeansppcenters}
\alias{kmeansppcenters}
\title{Find initial kmeans++ centroids.}
\usage{
kmeansppcenters(x, k)
}
\arguments{
\item{x}{Data matrix}

\item{k}{Number of clusters.}
}
\value{
A matrix containing the cluster centroids.
}
\description{
Find initial kmeans++ centroids.
}
\references{
Arthur & Vassilivitskii (2007) k-means++: the
advantages of careful seeding. In Proc SODA '07, 1027-1035.
}

