% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_data.R
\name{load_kdd_cup_1999_data}
\alias{load_kdd_cup_1999_data}
\title{Load the KDD Cup 1999 dataset}
\usage{
load_kdd_cup_1999_data(datapath)
}
\arguments{
\item{datapath}{The path to the data}
}
\value{
The data
}
\description{
Load the KDD Cup 1999 dataset
}

