% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{make_cfmat_string}
\alias{make_cfmat_string}
\title{Helper function to generate a string for a results table}
\usage{
make_cfmat_string(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A string
}
\description{
Helper function to generate a string for a results table
}
\keyword{internal}

