% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{make_experiment}
\alias{make_experiment}
\title{Run a core clustering experiment on a dataset.}
\usage{
make_experiment(dataset, clustering_func = NULL, sampling_func = NULL,
  alpha = 0.1, n_iter = 1000, method = "bootstrap", save = NULL,
  seed = 42, using_trim = FALSE)
}
\arguments{
\item{dataset}{A dataset: a list with (at least) the fields
\code{data} (the data as a matrix or data frame, depending on the
clustering function) and \code{classes} (the true, known classes of
the dataset).}

\item{clustering_func}{A clustering function to use, which given data returns cluster indices.}

\item{sampling_func}{A sampling function to use. Default is NULL. Only used when method is \code{distribution}.}

\item{alpha}{The confidenence level. Default 0.1}

\item{n_iter}{The number of iterations to use. The default is 1000.}

\item{method}{The method to use. String. Either \code{bootstrap} to
use bootstrapping or \code{distribution} to calculate the exact
co-occurrence probabilities.}

\item{save}{The full path to a file (rds-file) where to save the data.}

\item{seed}{Random seed. Default is NULL to not set the random seed.}

\item{using_trim}{Boolen. Does the clustering method use trimming?}
}
\value{
A list structure:
\describe{
\item{data}{The data}
\item{alpha}{The confidence level}
\item{iterations}{The number of iterations}
\item{cluster}{The initial clustering}
\item{cluster_core}{The core clustering}
\item{mat}{The co-occurrence probability matrix}
\item{amat}{The adjacency matrix}
\item{using_trim}{Whether or not the clustering method is using trimming}
\item{weak_cluster}{The index used to describe the weak clusters}
\item{classes_true}{The known true classes of the data items}
\item{clustering_function}{The used clustering function}
\item{time_start}{The timestamp when the experiment was started}
\item{time_stop}{The timestamp when the experiment was ready}
}
}
\description{
Run a core clustering experiment on a dataset.
}

