% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plots.R
\name{match_clusters}
\alias{match_clusters}
\title{Match two clusterings using the Hungarian algorithm.}
\usage{
match_clusters(n, cluster1, cluster2)
}
\arguments{
\item{n}{Number of items. Cluster indices shluld be in the range [1,n].}

\item{cluster1}{First vector of cluster indices.}

\item{cluster2}{Second vector of cluster indices.}
}
\value{
A vector of cluster indices for cluster1, so that they are best
matched to cluster2.
}
\description{
Match two clusterings using the Hungarian algorithm.
}

