% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{match_clusters_core}
\alias{match_clusters_core}
\title{Match core clusterings using the Hungarian algorithm.}
\usage{
match_clusters_core(res_1, res_2)
}
\arguments{
\item{res_1}{A result structure}

\item{res_2}{A result structure}
}
\value{
A list with the core clusters of res_1 and res_2 matched.
}
\description{
Match core clusterings using the Hungarian algorithm.
}

