% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plots.R
\name{plot_and_save_reslist}
\alias{plot_and_save_reslist}
\title{Plot (and optionally save) a list of result structures, one plot per figure.}
\usage{
plot_and_save_reslist(resl, basepath = NULL, plot_format = "pdf")
}
\arguments{
\item{resl}{A list of result structures from core clustering experiments}

\item{basepath}{The base path where to save the results.}

\item{plot_format}{The format of the plot. Default is \code{pdf}.}
}
\value{
Nothing
}
\description{
Plot (and optionally save) a list of result structures, one plot per figure.
}

