% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plots.R
\name{plot_data}
\alias{plot_data}
\title{Plot the data}
\usage{
plot_data(data, clusters = NULL, clusters_core = NULL, main = NULL)
}
\arguments{
\item{data}{The data}

\item{clusters}{The clusters}

\item{clusters_core}{The core clusters}

\item{main}{The title of the plot}
}
\value{
Nothing
}
\description{
Plot the data
}

