% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plots.R
\name{plot_reslist}
\alias{plot_reslist}
\title{Plot (and optionally save) a list of result structures, multiple plots in one figure.}
\usage{
plot_reslist(resl, save = FALSE, plot_format = "pdf", legend = FALSE)
}
\arguments{
\item{resl}{A list of result structures from core clustering experiments}

\item{save}{The filename where to save the figure}

\item{plot_format}{The format of the plot. Default is \code{pdf}.}

\item{legend}{Add the names of the used clustering functions to the plot}
}
\value{
Nothing
}
\description{
Plot (and optionally save) a list of result structures, multiple plots in one figure.
}

