% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plots.R
\name{plot_result}
\alias{plot_result}
\title{Plot (and optionally save) a result structure}
\usage{
plot_result(res, savename = NULL, main = NULL, plot_format = "pdf")
}
\arguments{
\item{res}{A result structure from a core clustering experiment}

\item{savename}{The filename where to save the figure}

\item{main}{The title of the plot}

\item{plot_format}{The format of the plot. Default is \code{pdf}.}
}
\value{
Nothing
}
\description{
Plot (and optionally save) a result structure
}

