% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{purity}
\alias{purity}
\title{Calculate the purity of a clustering}
\usage{
purity(cl_true, cl_other, ignore_class = NULL)
}
\arguments{
\item{cl_true}{The true classes of a dataset}

\item{cl_other}{A clustering results}

\item{ignore_class}{A class index to ignore, e.g., if trimming was used.}
}
\value{
A matrix with statistics.
}
\description{
Calculate the purity of a clustering
}

