% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{read_res}
\alias{read_res}
\title{Read a result structure from core clustering}
\usage{
read_res(classifier, dataset, base_path, is_list = TRUE, prefix = "res")
}
\arguments{
\item{classifier}{A classifier names}

\item{dataset}{The name of the dataset to read}

\item{base_path}{The base path to the results}

\item{is_list}{Is the result structure a list}

\item{prefix}{A prefix used in the filenames of the results}
}
\value{
A result structure from core clustering
}
\description{
Read a result structure from core clustering
}

