% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{read_result}
\alias{read_result}
\title{Read results from core clustering experiments stored in rds files.}
\usage{
read_result(pattern, path)
}
\arguments{
\item{pattern}{A pattern to match}

\item{path}{The path where the files are found}
}
\value{
A list with the fields \code{raw} containing the results as a matrix and \code{latex} containing the results as a latex table.
}
\description{
Read results from core clustering experiments stored in rds files.
}

