% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{read_result_set}
\alias{read_result_set}
\title{Read results from a filename and return the results as a matrix}
\usage{
read_result_set(f)
}
\arguments{
\item{f}{Full path to the file containing the results (an rds file).}
}
\value{
The results as a matrix.
}
\description{
Read results from a filename and return the results as a matrix
}

