% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_data.R
\name{read_uci_dataset}
\alias{read_uci_dataset}
\title{Read UCI datasets}
\usage{
read_uci_dataset(dslist, dataset_name, remove_class = TRUE, as_df = FALSE)
}
\arguments{
\item{dslist}{A list of available datasets obtained using the function \code{prepareDSList} from the package \code{readMLData}.}

\item{dataset_name}{The name of the dataset to read}

\item{remove_class}{Should the class be removed from the data. Default is TRUE.}

\item{as_df}{Should the data be returned as a data frame. Default is FALSE, which returns a matrix.}
}
\value{
A list with the fields:
\describe{
\item{data}{The data in the dataset}
\item{classes}{The classes in the dataset}
\item{name}{The name of the dataset}
}
}
\description{
Read UCI datasets
}

