% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{res_to_mat}
\alias{res_to_mat}
\title{Create a matrix of results given a result structure.}
\usage{
res_to_mat(res, alg_name)
}
\arguments{
\item{res}{A results structure}

\item{alg_name}{The name of the clustering algorithm}
}
\value{
A matrix
}
\description{
Create a matrix of results given a result structure.
}

