% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{special_table}
\alias{special_table}
\title{Create a results table in latex format given the name of a datset and a results structure.}
\usage{
special_table(dataset_name = "dataset_name", res)
}
\arguments{
\item{dataset_name}{The name of the dataset}

\item{res}{A results structure.}
}
\value{
Nothing.
}
\description{
Create a results table in latex format given the name of a datset and a results structure.
}

